Add error reporting capabilities
This required quite a big rewrite, which also triggered a change to use
Alternative instead of the old 'a grammar rule is a list of productions'
machinery of before. Having an Alternative instance for productions is
nice because the common functionality in 'many' and 'some' is at the
level of productions and not at the level of grammars as before. This is
achieved by just special-casing 'many' in productions --- not very
elegant, but extremely useful. It also means that the grammar language
is much closer to how most Haskell parser combinator libraries look
like.