Simplify parser externals and internals
Remove redundant type parameter to `Grammar`.  Remove separate handling
of nullables. It is handled correctly by the current machinery anyway.
This made it possible to get rid of the position fields in the `State`
and `Cont` types, as well as the list of null results in `Rule`, which
simplifies and speeds up the internals a bit.