Handle null-productions separately
... again.

Fix #14.

The problem was that the continuation of a nullable rule had the
potential to be followed before all occurrences of that rule had been
processed, leading to missing out on some continuations which means lost
results.

This commit brings back handling of null-productions similar to how it
was at an earlier point. Basically, we just handle such productions
separately.

Since they are handled separately we also attempt to remove
null-productions from the productions that are used in the "normal" code
path.